+++ Mlib/Physics/Advance_Times/Movables/Yaw_Pitch_Look_At_Nodes.cpp
@@ -65,7 +65,7 @@ void YawPitchLookAtNodes::set_absolute_model_matrix(const TransformationMatrix<f
             bullet_start_offset_,
             bullet_velocity_,
             gravity_,
-            1e-6,
+            (float)1e-6,
             10};
         t = aim.time;
         offset(1) = aim.aim_offset;
